finished menu on every responsivness
there is a need to try and make a more functionality with the buttons (hiding menu when clicked on something).
Then there is need to change gap between the sections and responsivness on the platform ( adding the "and (cursor: fine)" to the desktop @media method to make hoover desktop only. Next up there is a need to change a responsivness for the "Ourself" section

the last thing (mutual) is to change the forms based on schematics developed in this week